[PLAT-15830]Fix master_addresses for universe_join_existing_cluster subtask
Summary:
The master_address derived during the master_join_existing_cluster subtask is faulty since it does not use the node prefix.
Found while checking something else, where I deleted a pod and when it was re-created it was not able to resolve the
master addresses and failed.
Fixed by using masterAddress generated in parent task.

Test Plan: Manually verified

Reviewers: anijhawan, skurapati

Reviewed By: anijhawan, skurapati

Subscribers: anabaria, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39342