[BACKPORT 2024.2][PLAT-15848] Fix bug in logic for determining tables to verify bootstrap requirement
Summary:
The filter should be for tables which are *not* in `noSetupBootstrapRequiredTableUuids`. Prior to this diff, the negation was
missing.

Original commit: D39382

Test Plan:
Create an xCluster config.
Add some tables and databases to the source universe.
Edit the tables in replication and add the new tables to the config.
- Verify that only the tables which are `extra on source` are sent to
  verify bootstrap requirement.
{F303599}

Reviewers: rmadhavan, cwang, hzare

Reviewed By: rmadhavan, cwang

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39385