[BACKPORT 2.20][PLAT-15733] Fixing the initialization of schema versions
Summary:
If the initial set of schema version defined has a gap when we performed allSchemaSlice() we were not skipping the missing values and filling them in. This caused problems down the line when migrations that we thought had been run were not in fact run. This diff changes the allSchemasSlice() to read from migrations array and return sorted array of the keys.
Original commit: 45120c41fa2a833a14163e1edafc3e94b2dc57a0 / D39166
13ee5d743e5c014c9ce67af0db2b36595fa12542 / D39167

Test Plan: remove a key from migrations and ensure initial state has it missing

Reviewers: dshubin, sanketh

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39319