[BACKPORT 2024.2][#24365] YSQL: Run pg_cron jobs even on minute of failover
Summary:
Original commit: 876f1b9cb93564ed9b255b4c69a996e2c3aff7a3 / D38849
When the pg_cron leader moves due to LB, upgrade, or node crashes a new leader is elected after 60s `pg_cron_leader_lease_sec`. The new leader starts its execution from the current time, and skips the first minute. This leads to missed cron execution for a maximum of 2 minutes. Any job with a schedule to run in those minutes are missed.
Ex: If we have a job with schedule `0 4 * * *` (Run at 4:00 every day), and the cron leader fails over at 3:59.10 am, then we will not execute the 4am job.

This change addresses this by storing the time (in minutes) to the `pg_cron_leader` table every minute, and fetching it from the table the first time a new leader runs.
With this change order of events in pg_cron is:
1. If this is the first leader instance run then get the last minute stored in `pg_cron_leader` table. If there is nothing stored then returns the current minute.
1. Get the current time.
1. Compute which jobs need to be run since the last minute.
1. Log the run details in `cron.job_run_details` table.
1. Store the current minute in `pg_cron_leader` table.
1. Start the background worker for the new runs.

If we crash, or move the cron leader before step 4 then nothing is persisted, and the new leader will start from the previous stored last minute.
If we crash between 4 and 5, then we would have logged some job runs, but not started any workers yet. The new leader will mark the pending runs as 'failed' (with message 'server restarted') and retrigger new runs.
If we crash after 5, then the new leader will mark the pending runs as 'failed' and NOT retrigger new runs. Since the run could have completed but failed before it updated the `job_run_details` table we cannot retry.

This ensures that we at the least log a failed run and do not miss jobs silently. The user can monitor and handle these failed runs if needed.

If the universe is paused and resumed the clock would jump by a large amount. pg_cron natively handles this jump. If the duration is more than 3hrs, then the missed runs are skipped.

- Add `GetMemberAsUint` and `GetMemberAsUint64` helper functions to json_util.
- Updated `STATEFUL_SERVICE_IMPL_METHOD_HELPER` macro to include the `RpcContext` which has useful info like the deadline.
- Converted `PgCronLeaderService` to a RPC service.

**Upgrade/Rollback safety:**
`PgCronSetLastMinute` and `PgCronGetLastMinute` rpcs are added. These are used between pg, tserver, and the cron leader, all of which run on the same node. So these are upgrade and rollback safe.
Also, the feature is protected by `enable_pg_cron` preview flag.

Fixes #24365
Jira: DB-13277

Test Plan:
pg_cron-test
PgCronTest, ValidateStoredLastMinute
PgCronTest, PerMinuteTaskWithLeaderMove
PgCronTest, FailBeforeStoringLastMinute

Reviewers: tnayak, fizaa

Reviewed By: fizaa

Subscribers: ybase, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D39422