[PLAT-15824] Add universe detail page link to universe names in DR UI
Summary:
This diff is a UX improvement for the YBA disaster recovery UI.
We add a link to the universe details page for both the DR primary and the DR replica universe name so
the user can easily navigate to the participant universe pages.

Test Plan:
Create an xCluster DR config and verify that the user is able to navigate to the DR primary and DR replica universe
details page.
- Link color is black by default and orange on hover.
{F304173}
{F304174}

Reviewers: rmadhavan, kkannan, lsangappa, cwang, hzare

Reviewed By: cwang, hzare

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39441