[PLAT-13056] Allow text input of ssh private key for provider creation
Summary:
Currently the user must upload a file containing their SSH private key when creating a provider.
In some cases, it is much easier for the user to copy and paste the private key instead of uploading a file.

This diff exposes the ability to select which input type the user would like to use (file upload or text input).

Test Plan:
- Verify provider creation is successful with both types of ssh key input.
   - Required field validation should only be triggered on the input type which is selected.
      Tested by leaving both fields blank and verifying that only the input type selected gets validated for non-empty.

{F302506}
{F302507}

Reviewers: rmadhavan, kkannan, lsangappa

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39273