[#24560] YSQL: Make limit pushdown reset on top level read request
Summary:
When we push down the limit of a query to a request and we have to send a subsequent request, there is code in `pg_op.cc::PrepareNextRequest` to forget the pushed down limit and proceed with default row fetch settings. This is to avoid situations where an overly conservative limit pushdown does not hurt a query that requires fetching a lot more rows than the prescribed limit.

Before this change, this limit resetting mechanism did not reset the limit field of the appropriate read request during colocated index scans. The outermost request's limit is used by DocDB to decide how many rows to send over. Logic before this change reset the innermost limit instead of the outermost. This change fixes that fault.
Jira: DB-13595

Test Plan: ./yb_build.sh --java-test org.yb.pgsql.TestPgRegressColocation

Reviewers: telgersma, amartsinchyk

Reviewed By: telgersma, amartsinchyk

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D39453