[BACKPORT 2024.2][#24496] YSQL: Assume index contains 1000 rows if statistics not available
Summary:
Original commit: fe4811942b5340a87fda0d85028508708b2f49ef / D39297
If a table has not been ANALYZED, we do not know the number of rows it contains.
In this case, the `pg_class.reltuples` contains the value `0` in the YB versions
older than 2024.2 which were based on PG11. In the newer versions based on PG15,
`pg_class.relutples` contains the value `-1` if relutples is not known.

In YB, if reltuples of a table is not known, we assume the table has 1000 rows.
This helps the query planner make reasonable plan choices. In the case of index
however, this logic was missed. This can lead to incorrect index cost estimation
if the new cost model is used without running `ANALYZE` on the tables. The
old code model is not affected, because it does not utilize the number of
tuples in the index in estimating the cost.

This change fixes this issue, and likewise assumes that the index contains
1000 rows, if reltuples is not known. Tests are added which use the new
cost model without running ANALYZE on all or some of the tables used in the
query.
Jira: DB-13413

Test Plan: tbd

Reviewers: telgersma

Reviewed By: telgersma

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D39458