[#24574] docdb: TServers reject RBS requests for tablets they have deleted.
Summary:
This diff adds an in-memory map of tablets a tserver has processed delete tablet with tablet data state `DELETED` RPCs for to the `TSTabletManager`. Start remote bootstrap requests are checked against this map, and the tablet manager rejects requests to RBS a tablet it has previously deleted.

Note that delete tablet replica requests with table data state `TOMBSTONED` are not added to this map.
Jira: DB-13611

Test Plan:
```
% ./yb_build.sh release --cxx-test-filter-re remote_bootstrap-itest --cxx-test remote_bootstrap-itest --gtest_filter RemoteBootstrapITest.AcceptRBSAfterTabletTombstone
% ./yb_build.sh release --cxx-test-filter-re remote_bootstrap-itest --cxx-test remote_bootstrap-itest --gtest_filter RemoteBootstrapITest.RejectRBSAfterTabletDeletion
```

Reviewers: yyan, timur, bkolagani

Reviewed By: bkolagani

Subscribers: hsunder, slingam, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D39618