[#24671] YSQL: add postgres linter
Summary:
Add a basic linter for postgres code using [arc lint
script-and-regex][1].  Implement the script using a bash script that
runs some grep searches on the given file.  It is not powerful enough to
check for certain things such as whether indentation is aligned properly
with respect to other lines, but it is still better to have some basic
enforcement over nothing.

[1]: https://secure.phabricator.com/book/phabricator/article/arcanist_lint_script_and_regex/
Jira: DB-13743

Test Plan:
manually check arc lint on a few postgres files

Close: #24671

Reviewers: mihnea, tfoucher

Reviewed By: tfoucher

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D36535