[#24045] docdb: Switch test flag for persisted tserver registry to an auto flag.
Summary:
This switches the test flag for the persisted tserver registry to an auto flag.
We solve the problem of how to ensure tservers are persisted to the sys catalog when the flag is flipped on a running master leader by introducing a persisted bit on the proto.

There is code to write the tserver entry to the sys catalog on every heartbeat. Most heartbeats do not change any of the proto fields and so the write is skipped by the sys catalog machinery as being a no-op. This logic checks the dirty and clean protos of the COW object against each other. Because we set the persisted bit only when the flag is true, the dirty proto is different the first time we process a heartbeat after the flag is flipped. So the sys catalog write goes through.

For completeness, events that will persist a pre-registered tserver to the sys catalog after the flag is flipped are:
  - the tserver heartbeats
  - the tserver doesn't heartbeat for `tserver_unresponsive_timeout_ms` milliseconds, and is marked as unresponse by the CM background task

It is possible if a master leader crashes within `tserver_unresponsive_timeout_ms` milliseconds after the flag is flipped that not all tservers would be persisted to the sys catalog.
Jira: DB-12936

Test Plan:
```
./yb_build.sh --with-tests --cxx-test-filter-re master_heartbeat-itest --cxx-test master_heartbeat-itest  --gtest_filter 'PersistTSRegistryUpgradeTest.PersistTSRegistrationFlagFlip'
```

Reviewers: asrivastava

Reviewed By: asrivastava

Subscribers: yql, ybase, slingam

Differential Revision: https://phorge.dev.yugabyte.com/D38661