[BACKPORT 2024.2.0][PLAT-15882][PLAT-15883]Turn OIDC Auto create user on by default
Summary:
Original commit: D39494/19c8d7abe52e6cbbc64e15e50fc1d7492e9c024c

Based on discussions with Yu-Shen and Akki, from 2024.2, OIDC auto create users should be enabled by default, except for environments where OIDC was already setup previously.
This is because existing OIDC setups would have manually added users and the roles of these users would get overridden with the next login if auto create user is on.
So this feature is opt-in for existing OIDC setups and on by default for new setups.
So we make the default value of that key `true` in `reference.conf`.
The migration checks if OIDC was setup previously in the environment, in which case we set the value to false.

Also a minor bug fix to remove runtime config check from v2 group mappings APIs.

Test Plan:
Manually verified:
1. V2 group mappings APIs works with the runtime config disabled.
2. No-op incase auto create users is already setup.
3. Auto create user disabled if Oauth was setup previously.

Reviewers: #yba-api-review!, svarshney, agunta

Reviewed By: agunta

Subscribers: yugaware, yng, agunta

Differential Revision: https://phorge.dev.yugabyte.com/D39511