[PLAT-15516] DR failover task retry fails if PITR restore is retried
Summary: PITR restore is not idempotent at the DB level, and this diff implements logic in the PITR restore subtask such that if the restoration is previously, by a failed failover task, is created, it skips creating it again and only makes sure that the restoration is in RESTORED status.

Test Plan: Created a DR config and and manually failed the subtask right after the restoration is created and made sure a retry of the failover task reuses the already created restoration and succeeds.

Reviewers: cwang, vbansal, sanketh

Reviewed By: cwang

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39471