[PLAT-15928][xCluster][localProvider] Deflake testXClusterConfigHandleDroppedAndReAddedTable
Summary:
Previously, the new table that was created had a different schema than the dropped one and it could lead to issues if bootstrapping was skipped. It creates the new table with the same schema to make sure the test case runs successfully even if the bootstrapping is skipped.

The issue was observed at https://jenkins.dev.yugabyte.com/job/platform-phabricator-pipeline/33407/consoleFull.

Test Plan: Ran the test locally and it passed.

Reviewers: cwang, vbansal, sanketh

Reviewed By: cwang

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39551