[#23650] Fix testFastpathIntentdbSeeks failure in release builds
Summary:
Post 9c0de5d9, testFastpathIntentdbSeeks was failing in release builds
because Read Committed isolation is disabled by default in release
builds and hence the "error detail" that is returned with a 40001
serialization error is the following: "conflicts with same priority
transaction". When RC is enabled, the error detail is: "conflicts with higher
priority transaction". This is because the RC transaction in the test
uses kHighest priority or kHighPriTxnLowerBound - 1 depending on whether we
are using a debug or release build. This is because in release builds
`yb_enable_read_committed_isolation` is false, so the transaction falls
back to using Repeatable Read and hence honors the `yb_transaction_priority_lower_bound`
GUC that the test sets. In debug builds, `yb_enable_read_committed_isolation` is true, so
the transaction is actually a RC transaction and hence it ignores the GUC and uses kHighestPriority.

The assertion on error detail isn't adding any value to the test, so
this commit just removes it.

Test Plan: Jenkins

Reviewers: patnaik.balivada

Reviewed By: patnaik.balivada

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D39565