[#24703] Enhance pg15 upgrade template1 test
Summary:
There are a few database objects that can be created in template1. The existing test for upgrading template1 tested the presence of a function.

This diff expands on the existing test to add domains, extensions, event triggers, and schemas to template1, and validate that each behave as expected and that the relationships between them are as expected.

Test Plan:
```
./yb_build.sh --cxx-test pg15_upgrade-test --gtest_filter Pg15UpgradeTest.Template1
```

Reviewers: hsunder, fizaa, tfoucher

Reviewed By: hsunder

Subscribers: smishra, yql

Differential Revision: https://phorge.dev.yugabyte.com/D39542