[#22949] YSQL: Fix TestPgRegressPgTransactions
Summary:
### Context

YugabyteDB currently does not support DDL + DML transactions.

Take this transaction for example

```
1. BEGIN;
2. INSERT INTO TABLE foo ...
3. DROP TABLE foo;
4. CREATE TABLE foo;
5. INSERT TABLE foo ...
6. ROLLBACK
```

Each DDL (DROP/CREATE) is executed in a separate transaction and not the one started by BEGIN above. Let's call the latter, the outer transaction. Since the DROP happens in a separate transaction and DROP aborts all transactions that touched the table foo, the DROP transaction also aborts the outer transaction.

For this reason, typically, the INSERT in statement 5 above will see an abort. However, this is not a guarantee since the outer transaction only knows about the ABORT in the next transaction heartbeat. The timing may line up so that the INSERT goes through before the transaction hears about the abort.

### Issue

The test yb_pg_transactions.sql uses a similar pattern in one of the tests.

```
BEGIN;
	savepoint x;
		CREATE TABLE koju (a INT UNIQUE);
		INSERT INTO koju VALUES (1);
		INSERT INTO koju VALUES (1);
	rollback to x;
	DROP TABLE koju; -- YB: above CREATE TABLE is not rolled back due to lack of transactional DDL

	CREATE TABLE koju (a INT UNIQUE);
\set VERBOSITY terse \\ -- YB: suppress "expired or aborted by a conflict" message because it prints a file line number
	INSERT INTO koju VALUES (1); -- YB: TODO(#22949): fix output
\set VERBOSITY default \\ -- YB
	INSERT INTO koju VALUES (1);
ROLLBACK;
```

This makes the output on the INSERT that has YB_TODO non-deterministic.

This is a new test introduced as part of pg15 merge in 475dbd9f0358 / D36033.

### Changes

Remove the DROP/CREATE statement from the transaction. This makes the output deterministic.
Jira: DB-11863

Test Plan:
Jenkins: test regex: .*TestPgRegressPgTransactions.*

Flaky test is not reproducible locally. Tested by running the same test for 10+ times on detective.

Reviewers: jason, bkolagani, smishra

Reviewed By: jason

Subscribers: yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D39668