[#24719] YSQL: Support ysql major upgrade on auth enabled universes
Summary:
In non-test universes secure connections and authentications are always enabled. In these universes we have to use the local socket to connect to YSQL.
This change allows us to pass the socket addresses for the old and new clusters to `pg_upgrade`.

If we do not find a local yb-tserver on the master leader node then we fallback to the host connection. This makes sure universe that do not place master and tservers on the same node like Kubernetes deployments to still function if they do not have auth enabled. A workaround for these universe is to add the password to `~/.pgpass` file on the master nodes.

- Updated `GetClosestLiveTserver` to track if a local tserver was picked.
- Removed the `socketdir` option from pg_upgrade and replaced it with `old-socketdir`/`s` and `new-socketdir`/`S`

Fixes #24719
Jira: DB-13793

Test Plan:
MasterTest.TestGetClosestLiveTserver
Pg15UpgradeTest.NoTserverOnMasterNode
Pg15UpgradeTestWithAuth.UpgradeAuthEnabledUniverse
Pg15UpgradeTestWithAuth.NoTserverOnMasterNode

Reviewers: telgersma, fizaa, tfoucher

Reviewed By: telgersma

Subscribers: sanketh, yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D39566