[#24597] YSQL: Disable catalog version check during upgrade
Summary:
### Background

Currently, major version upgrade fails with the following error on a cluster with global breaking DDL changes:
```
2024-10-30 15:05:25.338 EDT [50162] FATAL:  The catalog snapshot used for this transaction has been invalidated: expected: 2, got: 1: MISMATCHED_SCHEMA
2024-10-30 15:05:25.338 EDT [50162] STATEMENT:
	UPDATE pg_class   SET relacl = (SELECT array_agg(a.acl) FROM  (SELECT E'=r/"postgres"' as acl   UNION SELECT unnest(pg_catalog.acldefault(    CASE WHEN relkind = 'S' THEN 's'          ELSE 'r' END::"char",10::oid)) ) as a)   WHERE relkind IN ('r', 'v', 'm', 'S')  AND relacl IS NULL;

```

During upgrade, initdb executes some queries on the new cluster after the initial bootstrap is complete.
The backend that executes these queries internally sets its catalog version using `YBGetCatalogCacheVersion`.
This read gets routed to the PG15 copy of the catalog version table, which is initialized with a value of 1.
However, when a request is sent from this backend, it is evaluated against the PG11 catalog version, as during
a major version upgrade, the master provides the PG11 catalog version to the tservers via the heartbeat mechanism.
This discrepancy results in a catalog version mismatch.

### Solution

Fix the problem by disabling catalog version checks during upgrade. DDLs that execute during the pg_restore phase
do not increment the catalog version, so we have no reason to perform a check.
We may have to make this change more fine-grained in the future, depending on what DDLs we white-list during mixed mode
and how we go about it.
Jira: DB-13635

Test Plan: ./yb_build.sh release --cxx-test pg15_upgrade-test --gtest_filter Pg15UpgradeTest.GlobalBreakingDDL

Reviewers: hsunder, tfoucher, telgersma

Reviewed By: telgersma

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D39622