[BACKPORT 2.23.0.3690][PLAT-15952] cgroup values not getting updated on resize
Summary:
Original commit: None / D39682
we all assumed this was checked once for the whole block. But, it is checked before ansible executes
every task in the block.
What this means is we do the first set_fact that will set systemd_dir to /etc/systemd/system. Now that
it is set, we drop out of the rest of the block because the where clause is no longer true

we now will always try to set systemd_dir, like other playbooks

Test Plan:
did a resize with cgroup size change. Validated correct cgroup service file gets updated, along with
the related sysfs cgroup file.

Reviewers: daniel, nbhatia, dnolan, svarshney, tsinghania

Reviewed By: svarshney, tsinghania

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39684