[BACKPORT 2.20][PLAT-15273]Remote files download bug fix
Summary:
Original commit: 7ece03553c0e848dee28d0c5ca4ba43c1ceec365/D39722

This diff has the fix for a bug where we assume the same tmp directory for the YBA and the db nodes.
As a result support bundle creation fails when we set `yb.filepaths.TmpDirectory` to any path which is absent on the db nodes.

Fix for both the python script and the node agent flow.

Test Plan: Manually verified that support bundle creation succeeds when the tmp directory is changed on either YBA or the db nodes via gflags.

Reviewers: nsingh, skurapati

Reviewed By: nsingh

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39759