[#24823] DocDB: Fix use after free bugs in MasterPathHandlers::HandleTablePage[JSON]
Summary:
These routines have the code:
```
~/code/yugabyte-db/src/yb/master/master-path-handlers.cc:2113:
    auto sorted_locations = TabletReplicaMapToSortedVector(*tablet->GetReplicaLocations());
    auto l = tablet->LockForRead();
```

This is an error because the lifetime of the result returned by GetReplicaLocations is only as long as the shared_ptr is returned is held but the code above discards that shared_ptr immediately.

Fixes #24823

The shared_ptr needs to be held until we no longer need to use the locations.
Jira: DB-13927

Test Plan:
On Jenkins, this failed with:
```
ybd tsan --cxx-test integration-tests_master_path_handlers-itest --gtest_filter MasterPathHandlersItest.ShowDeletedTablets --clang17
```
I am sadly unable to get this to reproduce on my dev server even with tens of tries.

Reviewers: jhe

Reviewed By: jhe

Subscribers: ybase, esheng

Differential Revision: https://phorge.dev.yugabyte.com/D39787