[BACKPORT 2.20][#24663] YSQL: Fix foreign key on unique index with different column orders
Summary:
For a foreign key constraint on the unique index of the referenced relation, the function YBCBuildYBTupleIdDescriptor() incorrectly assumes that the order of columns in foreign and unique key constraints would be the same. This leads to incorrect foreign key violation errors and data inconsistency issues when that is not the case (see the added test). Fix it by using the correct attribute number in the index relation.
Jira: DB-13730

Original commit: f99d41331aff5c43c5a80878dca16b846719cbb1 / D39538

Test Plan:
./yb_build.sh --java-test org.yb.pgsql.TestPgForeignKey
./yb_build.sh --java-test org.yb.pgsql.TestPgForeignKeyBatching
./yb_build.sh --java-test org.yb.pgsql.TestPgRegressForeignKey

Reviewers: dmitry, devansh.saxena

Reviewed By: devansh.saxena

Subscribers: yql, smishra

Differential Revision: https://phorge.dev.yugabyte.com/D39776