[#24772] docdb: Fix minicluster restart to work with a persisted ts registry.
Summary:
`WaitForTabletServerCount` checks the master leader's `TSManager` class to see if it is aware of all the tservers in the test universe. The `RestartSync` method uses `WaitForTabletServerCount` to block tests until the master leader has heard from all tservers in the test cluster. However with the persisted ts registry this method is broken, as a master leader loads after a restart with descriptors for all tservers in memory. There is some important in-memory state that is not persisted, such as tablet replicas, which can break some tests.

This diff adds a check that the master leader has a tablet report from each tablet server, fixing `WaitForTabletServerCount`.
Jira: DB-13862

Test Plan:
```
./yb_build.sh --cxx-test-filter-re snapshot-test --cxx-test snapshot-test --gtest_filter SnapshotTest.HideTablesCoveredBySnapshot -n 10
```

Reviewers: asrivastava

Reviewed By: asrivastava

Subscribers: ybase, slingam

Differential Revision: https://phorge.dev.yugabyte.com/D39782