[#24362] Revert: xCluster: Raise NOTICE on creation of index of bi-directional replication
Summary:
This reverts commit e9b53aee21f1441c0c1b90fc019fd8d6c0d2ca25.

Because of the way ysql and master handles the create table retires it is not guaranteed to return the message every time.
Also there is no easy way to track when the index create gets stuck waiting on xcluster or the backfill. The best way to check this today is to look at tasks running in master which already has a waiting for bidirectional xcluster task.

We need a better way to track index backfill progress in the YSQL client, which will handle xCluster as well in future.
Jira: DB-13275

Test Plan: Jenkins

Reviewers: jhe, slingam, amartsinchyk, xCluster

Reviewed By: jhe

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D39802