[#24727]: Skip index build during YSQL major version upgrade
Summary:
### Background

In the standard PG upgrade process, the steps are as follows:

1. `pg_restore` is initiated through `create_new_objects()`.
2. Cluster objects, including indexes, are created: during index creation, an internal index build is triggered, but no data is inserted into the index since the base tables are empty at that stage. The `index_build()` function marks the relation as having an index by setting `relhasindex`.
3. After `create_new_objects()` completes, `transfer_new_all_tablespaces()` transfers data for both indexes and tables from the old cluster's relfilenodes to the new cluster's relfilenodes.

In the context of a YSQL major version upgrade, the process diverges as follows:

1. `pg_restore` is invoked via `create_new_objects()`.
2. When a relation is created, the corresponding DocDB table is linked on the master leader as part of the `CREATE` command itself.
3. Therefore, when a `CREATE INDEX` command is being executed during the restore, the base table is already populated, and the DocDB table for the index is linked before the index build occurs. Consequently, during the index build in YB, data is re-inserted into the index, resulting in a duplicate build. This can lead to upgrade failures due to constraint violation errors when unique indexes are involved.

### Solution

To resolve this issue, block the index build in `ybcinbuild`, while still allowing other logic in `index_build` (such as `index_update_stats`) to execute. This ensures that the `relhasindex` flag is correctly set to true on the base table.

Note: Following this fix, the `reltuples` for an index will be cleared after an upgrade. This aligns with PG behaviour and our current behaviour for regular tables.

Test Plan: ./yb_build.sh release --cxx-test pg15_upgrade-test --gtest_filter Pg15UpgradeTest.Indexes

Reviewers: tfoucher, telgersma

Reviewed By: tfoucher

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D39784