[PLAT-15943] Yba Installer symlinking to be idempotent
Summary:
We found that the symlink function in yba-installer does not handle
the case where the symlink was already created. A new symlink in this
scenario leads to a file system loop.

We now detect when the symlink already exists, and delete and recreate it
if it does.

If the target file exists and is not a symlink, we should fail, as this
is an unexpected scenario

Test Plan: tested new symlink function during migration`

Reviewers: muthu, sanketh

Reviewed By: muthu

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39748