[PLAT-16052]Fix SSO login and improve logging for easier debugging
Summary:
The JWT library was updated with commit D39576.
The newer version had breaking changes due to which SSO login on no longer works on master.
Revert the change and improve logging for easier debugging.

Test Plan: Manually verified that SSO was breaking on master and works after the revert.

Reviewers: svarshney

Reviewed By: svarshney

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D39825