[BACKPORT 2024.1.3][#20321] docdb: Fix move transaction status RPC cleanup race with yb::Transaction destructor
Summary:
It is currently possible for move transaction status RPC to finish after `yb::Transaction`
destructs, because the handler only uses a weak_ptr. In this case, we do not clean up the RPC
handle, because we no longer have access to the `Rpcs` object or the handle.

This causes promotion tests that run into this case to occasionally timeout, because the
`Rpcs::Abort` call in the `yb::Transaction` destructor waits for the move transaction status RPCs
to be unregistered, but this never happens since the weak_ptr lock fails (due to having entered
the destructor).

This diff changes the move transaction status code path to maintain strong references to the
transaction object instead, so that the RPC cleanup can happen.
Jira: DB-9305

Original commit: 71412cfcffdb263627ec941b2a423964570a8525 / D38792

Test Plan:
Jenkins. Also ran GeoTransactionsPromotionTest.TestPromotionReturningToAbortedState 500x
and ensured timeout did not happen.

Reviewers: sergei

Reviewed By: sergei

Subscribers: ybase, rthallam

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D39829